rag_tools/requirements.txt
# Runtime dependencies only.
# Recommended Python version: 3.10 - 3.12

# Core data handling
numpy>=1.26,<3
pandas>=2.1,<3
requests>=2.31,<3

# Retrieval and embeddings
faiss-cpu>=1.8,<2
sentence-transformers>=3,<6

# LLM clients
openai>=1.30,<3

# Crawling and parsing
beautifulsoup4>=4.12,<5
selenium>=4.20,<5
webdriver_manager>=4,<5
lxml>=5,<7

# Utilities
tqdm>=4.66,<5
python-dotenv>=1,<2
PyYAML>=6,<7
